feat(base-payment-lib): Move `GmsBottomSheetDialogFragment` to `internal-payment-sdk`
EC-137
