+++ gini-internal-payment-sdk/library/src/main/java/net/gini/android/internal/payment/utils/extensions/View.kt
@@ -4,9 +4,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
 
 internal fun View.getLayoutInflaterWithGiniPaymentTheme(): LayoutInflater =
     LayoutInflater.from(context.wrappedWithGiniPaymentTheme())
+
+internal fun Fragment.getLayoutInflaterWithGiniPaymentTheme(inflater: LayoutInflater): LayoutInflater {
+    return inflater.cloneInContext(requireContext().wrappedWithGiniPaymentTheme())
+}
 
 internal fun View.hideKeyboard() {
     ContextCompat.getSystemService(context, InputMethodManager::class.java)?.let { imm ->
