+++ merchant-sdk/sdk/src/main/java/net/gini/android/merchant/sdk/paymentComponentBottomSheet/PaymentComponentBottomSheet.kt
@@ -14,7 +14,7 @@ import net.gini.android.internal.payment.paymentComponent.PaymentComponent
 import net.gini.android.merchant.sdk.databinding.GmsBottomSheetPaymentComponentBinding
 import net.gini.android.merchant.sdk.paymentcomponent.PaymentComponentView
 import net.gini.android.merchant.sdk.util.BackListener
-import net.gini.android.merchant.sdk.util.GmsBottomSheetDialogFragment
+import net.gini.android.internal.payment.utils.GpsBottomSheetDialogFragment
 import net.gini.android.merchant.sdk.util.autoCleared
 import net.gini.android.merchant.sdk.util.extensions.setBackListener
 import org.jetbrains.annotations.VisibleForTesting
@@ -23,7 +23,7 @@ internal class PaymentComponentBottomSheet private constructor(
     paymentComponent: PaymentComponent?,
     reviewFragmentShown: Boolean,
     backListener: BackListener? = null
-): GmsBottomSheetDialogFragment() {
+): GpsBottomSheetDialogFragment() {
     constructor(): this(null, false)
 
     private var binding: GmsBottomSheetPaymentComponentBinding by autoCleared()
